edit_update.sh

+++ miyagawpeda_update.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
-rm -rf ~/peda/miyagawpeda.py
+if test -e ~/peda/miyagawpeda.py ;then
+	rm -rf ~/peda/miyagawpeda.py
+fi
 #git clone https://github.com/miyagaw61/miyagawpeda.git ~/peda/miyagawpeda
 line=$(cat ~/peda/peda.py | grep -n "def xprint" | sed -E "s/(.*):.*:/\1/g")
 head -n $((${line} - 1)) ~/peda/peda.py | tail -n $((${line} - 1)) > ~/peda/miyagawpeda/peda.before.txt
